+++ src/client/src/System/GameStateMovePlanet.cpp
@@ -241,7 +241,7 @@ bool    CGameStateMovePlanet::On_WM_MOUSEWHEEL(WPARAM wParam, LPARAM lParam) {
 #ifdef _DEBUG
   if( g_GameDATA.m_bObserverCameraMode )
   {
-    ObserverCameraZoomInOut( (float)-(zDelta) );
+    ObserverCameraZoomInOut( (int)((float)-(zDelta)) );
   }
 #endif
