+++ src/client/src/Util/CFileSystemNormal.cpp
@@ -255,7 +255,7 @@ void CFileSystemNormal::WritePascalString(const char* pStr) {
   if ( pStr == nullptr || m_pFile == nullptr )
     return;
 
-  short nLength = strlen( pStr );
+  short nLength = (short)strlen( pStr );
 
   /// 두바이트로 길이를 표현한다.
   if ( nLength > 0x7F ) {
